# Bug 2014497: Optimize name validation regex for forms