--- a/frontend/packages/console-shared/src/utils/__tests__/yup-validations.spec.ts
+++ b/frontend/packages/console-shared/src/utils/__tests__/yup-validations.spec.ts
@@ -1,0 +1,22 @@
+import { nameRegex } from '../yup-validations';
+
+describe('nameRegex', () => {
+  it('should match valid resource names', () => {
+    const validNames = ['appname', 'app-name', 'app-name123', 'app-name123-app'];
+    validNames.forEach((name) => expect(name).toMatch(nameRegex));
+  });
+
+  it('should not match invalid resource names', () => {
+    const invalidNames = [
+      'AppName',
+      '4appname',
+      '-app-name',
+      'app$name!',
+      'app name',
+      'app--name',
+      '',
+      '-',
+    ];
+    invalidNames.forEach((name) => expect(name).not.toMatch(nameRegex));
+  });
+});