--- a/frontend/packages/console-shared/src/utils/yup-validations.ts
+++ b/frontend/packages/console-shared/src/utils/yup-validations.ts
@@ -1,7 +1,7 @@
 import { TFunction } from 'i18next';
 import * as yup from 'yup';
 
-export const nameRegex = /^([a-z]([-a-z0-9]*[a-z0-9])?)*$/;
+export const nameRegex = /^[a-z]([a-z0-9]-?)*[a-z0-9]$/;
 export const nameValidationSchema = (t: TFunction, maxLength = 263) =>
   yup
     .string()