**Fixes**: 
https://bugzilla.redhat.com/show_bug.cgi?id=2014497
<!-- For e.g Fixes: https://issues.redhat.com/browse/ODC-XXX -->

**Analysis / Root cause**: 
The regex for resource name validation was very slow and would often result in recursion errors in the browser console, causing the name to not be validated and the formik validation to crash.
<!-- Briefly describe analysis of US/Task or root cause of Defect -->

**Solution Description**: 
Removed the capture groups from the regex, since the captured strings [weren't used](https://regex101.com/r/jsDbli/2) anyway, and there was a lot of repeated iteration due to the nested group. Also cleaned up the regex to avoid recursion.

I wrote a quick function for checking the time taken by both the old and new regexes to validate a given string, in the browser console:
```js
let checkTime = str => {
  let r1 = /^([a-z]([-a-z0-9]*[a-z0-9])?)*$/;
  console.time('test-old');
  r1.test(str);
  console.timeEnd('test-old');

  let r2 = /^[a-z][-a-z0-9]*[a-z0-9]$/;
  console.time('test-new');
  r2.test(str);
  console.timeEnd('test-new');
}
```
When checking a valid string, there was not much of a noticeable difference.
For the string `app-name-test-test0-test1-test2-test3-test4-test5`:

Browser|old|new
-|-|-
Firefox|0ms|1ms
Chromium|0.046ms|0.032ms

But when checking for an _invalid_ string, there was a surprisingly large time difference between the two!
For the string `app-name-test-test0-test1-test2-test3-invaliD-!@#`:

Browser|old|new
-|-|-
Firefox|767ms|1ms
Chromium|726.21ms|0.028ms

My guess is that all the time is being spent in recursively capturing the last valid sub group from the string, while in the first case, it simply captures the whole string due to it being fully valid.

**Unit test coverage report**: 
Added tests for regex check.
<!-- Attach test coverage report -->

**Browser conformance**: 
<!-- To mark tested browsers, use [x] -->
- [x] Chrome
- [x] Firefox
- [ ] Safari
- [ ] Edge

/kind bug