@rottencandy 2021-11-26
> Can you please also remove the unnecessary try-catch block in `createComponentName` in `packages/dev-console/src/components/import/import-validation-utils.ts` introduced in #10013

Removed.