@rottencandy 2021-11-26
> `node--test` is not a valid string for the name. It would be nice if you include it in regex as well.

@vikram-raj updated. There's a 0.01ms increase in match time now. I think it's not too bad.