@rottencandy 2021-11-29
/cherrypick release-4.9